Moved AnomalyRegion.py => anomaly_region.py
