+++ src/nupic/regions/anomaly_region.py
